Move exporter to file dir
